+++ stokApp/package.json
@@ -14,6 +14,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^5.3.0",
+    "react-router-dom": "^7.0.1",
     "react-select": "^5.8.3",
     "react-select-tailwind": "^1.0.2"
   },
