+++ stokApp/src/components/GraphWidget.jsx
@@ -14,15 +14,16 @@ export default function GraphWidget({ symbol }) {
           "autosize": true,
           "symbol": "${symbol}",
           "interval": "D",
-          "timezone": "Etc/UTC",
+          "timezone": "America/New_York",
           "theme": "dark",
           "style": "1",
           "locale": "en",
           "withdateranges": true,
-          "height": "400",
+          "height": "450",
           "backgroundColor": "rgba(53, 28, 117, 1)",
           "gridColor": "rgba(255, 255, 255, 0.06)",
-          "allow_symbol_change": false,
+          "allow_symbol_change": true,
+          "details": true,
           "calendar": false,
           "support_host": "https://www.tradingview.com"
         }`;
