+++ stokApp/src/pages/AccountSettingsPage.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const AccountSettingsPage = () => {
+  return (
+    <div className="p-5">
+      <h1 className="text-3xl text-violet-500">Account Settings</h1>
+      <p>Manage your account details here.</p>
+    </div>
+  );
+};
+
+export default AccountSettingsPage;
