+++ stokApp/src/pages/LoginPage.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const LoginPage = () => {
+  return (
+    <div className="flex justify-center items-center min-h-screen">
+      <h1 className="text-3xl text-violet-500">Welcome to SimplyStocks - Login Page</h1>
+    </div>
+  );
+};
+
+export default LoginPage;
