+++ stokApp/src/pages/PortfolioPage.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const PortfolioPage = () => {
+  return (
+    <div className="p-5">
+      <h1 className="text-3xl text-violet-500">Portfolio</h1>
+      <p>Manage your stock portfolio here.</p>
+    </div>
+  );
+};
+
+export default PortfolioPage;
